refactor: extract shared CDP helpers for E2E tests
- findPageTarget, evalInObsidian을 cdp-helpers.mjs로 추출
- run-e2e.mjs, run-settings-e2e.mjs 모두 공유 모듈 import
- 51줄 중복 코드 제거
