+++ tests/e2e/run-e2e.mjs
@@ -17,11 +17,12 @@
  *   CDP_TARGET_ID=<id> node tests/e2e/run-e2e.mjs       # specify CDP page target
  */
 
+import { findPageTarget, evalInObsidian } from './cdp-helpers.mjs';
+
 // ---------------------------------------------------------------------------
 // Config
 // ---------------------------------------------------------------------------
 
-const CDP_PORT = process.env.CDP_PORT || 9222;
 const PLUGIN_ID = 'auto-note-importer';
 const MULTI_CONFIG_TABLE_ID = 'tblZO35AeSdSmI3rr';
 const MULTI_CONFIG_FOLDER = 'E2E-Multi';
@@ -30,50 +31,6 @@ const MULTI_CONFIG_ID = 'e2e-cfg-2';
 // Test records created during setup — deleted during cleanup
 let testRecordIds = [];
 let multiConfigRecordIds = [];
-
-// ---------------------------------------------------------------------------
-// CDP helpers
-// ---------------------------------------------------------------------------
-
-async function findPageTarget() {
-  const override = process.env.CDP_TARGET_ID;
-  if (override) return override;
-
-  const resp = await fetch(`http://localhost:${CDP_PORT}/json/list`);
-  const targets = await resp.json();
-  const page = targets.find(t => t.type === 'page' && t.url.includes('obsidian'));
-  if (!page) throw new Error('No Obsidian page target found. Is Obsidian running with --remote-debugging-port?');
-  return page.id;
-}
-
-function evalInObsidian(targetId, expression, timeout = 20000) {
-  const wsUrl = `ws://localhost:${CDP_PORT}/devtools/page/${targetId}`;
-  return new Promise((resolve, reject) => {
-    const timer = setTimeout(() => reject(new Error(`Timeout (${timeout}ms)`)), timeout);
-    const ws = new WebSocket(wsUrl);
-    ws.addEventListener('open', () => {
-      ws.send(JSON.stringify({
-        id: 1,
-        method: 'Runtime.evaluate',
-        params: { expression, awaitPromise: true, returnByValue: true }
-      }));
-    });
-    ws.addEventListener('message', (e) => {
-      const result = JSON.parse(e.data);
-      if (result.id === 1) {
-        clearTimeout(timer);
-        if (result.result?.exceptionDetails) {
-          resolve({ __error: result.result.exceptionDetails.exception?.description || 'Unknown error' });
-        } else {
-          try { resolve(JSON.parse(result.result.result.value)); }
-          catch { resolve(result.result.result.value); }
-        }
-        ws.close();
-      }
-    });
-    ws.addEventListener('error', (err) => { clearTimeout(timer); ws.close(); reject(err); });
-  });
-}
 
 // ---------------------------------------------------------------------------
 // Obsidian-side helpers (injected into eval expressions)
